+++ client/.changeset/great-carrots-behave.md
@@ -1,0 +1,5 @@
+---
+"@wikijump/util": minor
+---
+
+Added CSS template string function.
